Remove call to expand QS panel while pressing the Airplane QS tile
Currently if you bring down the QS panel a quarter of the way, just enough where you
can see the QS tiles and press on the Airplane QS tile it expands the QS panel.

This is not stock behavior.

The call was added in the commit below but I don't believe that it was intended to ever
be there because it doesn't make sense. If you're on a secured lockscreen and bring down
the QS panel, press the QS tile and unlock your device, the QS panel is already down. There
is really no need to call it again.

https://github.com/DirtyUnicorns/android_frameworks_base/commit/9774af33cd85ef976f9b0e4f32c8ab5f1d57480a

Change-Id: I4679d0a637dec60df2d49033e267de35cae4d758
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>
Signed-off-by: Srijith Bharadwaj <srijithdamerla2001@gmail.com>