QS: add right/left padding even if it's 0 for collapsed statusbar
For devices without a rounded display we are setting rounded_corner_content_padding
to 0 to avoid useless left/right padding in the statusbar.
But this looks odd in the expanded statusbar / qs panel because other elements
like the brightness slider buttons or the date have the padding anyway.

We can handle those cases separately.

Thanks anirudhgupta109 for the suggestion.

SKULSHADY: Change it to 12dp to align it with notification content.

Change-Id: Icce817f11423060f3691e3350d0900fa7247b942
Signed-off-by: Srijith Bharadwaj <srijithdamerla2001@gmail.com>